+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* [PR-91](https://github.com/itk-dev/economics/pull/91)
+  Updated standard price on clients
 * [PR-89](https://github.com/itk-dev/economics/pull/89)
   Cleaned up Twig templates.
 * [PR-88](https://github.com/itk-dev/economics/pull/88)